Reduced double calculations

+++ include/fft.h
@@ -1,3 +1,3 @@
 #define N_POINT 1024
 
-void fft_Npt(int *input, double *X_out, double *Y_out,int N);
+void fft_Npt(int *input, int *X_out, int *Y_out,int N);
